move pkg integration test outside
