+++ internal/api/e2e/api/globaldatasource_test.go
@@ -13,7 +13,7 @@
 
 //go:build integration
 
-package e2e
+package api
 
 import (
 	"testing"
