+++ internal/api/e2e/client/test_utils.go
@@ -13,7 +13,7 @@
 
 //go:build integration
 
-package v1
+package client
 
 import (
 	"net/http/httptest"
@@ -21,11 +21,12 @@ import (
 
 	e2eframework "github.com/perses/perses/internal/api/e2e/framework"
 	"github.com/perses/perses/internal/api/shared/dependency"
+	"github.com/perses/perses/pkg/client/api/v1"
 	"github.com/perses/perses/pkg/client/perseshttp"
 	"github.com/perses/perses/pkg/model/api"
 )
 
-func withClient(t *testing.T, testFunc func(ClientInterface, dependency.PersistenceManager) []api.Entity) {
+func withClient(t *testing.T, testFunc func(v1.ClientInterface, dependency.PersistenceManager) []api.Entity) {
 	server, _, persistenceManager := e2eframework.CreateServer(t)
 	defer server.Close()
 	persesClient := createClient(t, server)
@@ -34,12 +35,12 @@ func withClient(t *testing.T, testFunc func(ClientInterface, dependency.Persiste
 
 }
 
-func createClient(t *testing.T, server *httptest.Server) ClientInterface {
+func createClient(t *testing.T, server *httptest.Server) v1.ClientInterface {
 	restClient, err := perseshttp.NewFromConfig(perseshttp.RestConfigClient{
 		URL: server.URL,
 	})
 	if err != nil {
 		t.Fatal(err)
 	}
-	return NewWithClient(restClient)
+	return v1.NewWithClient(restClient)
 }
